+++ src/main/scala/com/cloudray/scalapress/search/SearchResult.scala
@@ -3,7 +3,7 @@ package com.cloudray.scalapress.search
 /** @author Stephen Samuel */
 case class SearchResult(refs: Seq[ItemRef] = Nil,
                         facets: Seq[Facet] = Nil,
-                        count: Long = 0)
+                        count: Int = 0)
 
 case class ItemRef(id: Long,
                    itemType: Long,
